+++ src/clj_kondo/impl/analyzer.clj
@@ -1559,9 +1559,10 @@
   (swap! (:namespaces ctx) assoc-in [base-lang lang current-ns :gen-class] true)
   nil)
 
-(defn analyze-reify [ctx expr]
-  (let [children (next (:children expr))
-        children (map (fn [node]
+(defn analyze-extend-type-children
+  "Used for analyzing children of extend-type, reify and extend-protocol."
+  [ctx children]
+  (let [children (map (fn [node]
                         (if (= :list (tag node))
                           (update node :children (fn [children]
                                                    (cons (utils/token-node 'clojure.core/fn)
@@ -1569,6 +1570,24 @@
                           node))
                       children)]
     (analyze-children ctx children)))
+
+(defn analyze-reify [ctx expr]
+  (let [children (next (:children expr))]
+    (analyze-extend-type-children ctx children)))
+
+(defn analyze-extend-type [ctx expr]
+  (let [children (next (:children expr))
+        ctx (if (identical? :cljs (:lang ctx))
+              (update-in ctx [:config :linters :unresolved-symbol :exclude]
+                         (fn [config]
+                           (conj config
+                                 'number 'function 'default 'object 'string)))
+              ctx)]
+    (analyze-extend-type-children ctx children)))
+
+(defn analyze-cljs-exists? [ctx expr]
+  (run! #(analyze-usages2 (ctx-with-linters-disabled ctx [:unresolved-symbol :unresolved-namespace]) %)
+        (next (:children expr))))
 
 (defn analyze-call
   [{:keys [:top-level? :base-lang :lang :ns :config :dependencies] :as ctx}
@@ -1744,14 +1763,13 @@
                       doto
                       (analyze-expression** ctx (macroexpand/expand-doto ctx expr))
                       reify (analyze-reify ctx expr)
-                      (. .. proxy extend-protocol
-                         defcurried extend-type specify!)
+                      (extend-protocol extend-type specify!) (analyze-extend-type ctx expr)
+                      (. .. proxy defcurried)
                       ;; don't lint calls in these expressions, only register them as used vars
                       (analyze-children (ctx-with-linters-disabled ctx [:invalid-arity
                                                                         :unresolved-symbol
                                                                         :type-mismatch])
                                         children)
-                      ;; TODO: extend-type + specify! are similar in CLJS
                       (proxy-super)
                       (analyze-proxy-super ctx expr)
                       (amap)
@@ -1802,6 +1820,7 @@
                       (analyze-hof ctx expr resolved-as-name)
                       (ns-unmap) (analyze-ns-unmap ctx base-lang lang ns-name expr)
                       (gen-class) (analyze-gen-class ctx expr base-lang lang ns-name)
+                      (exists?) (analyze-cljs-exists? ctx expr)
                       ;; catch-all
                       (case [resolved-as-namespace resolved-as-name]
                         [clj-kondo.lint-as def-catch-all]