+++ test/clj_kondo/main_test.clj
@@ -1743,12 +1743,27 @@ foo/foo ;; this does use the private var
  :methods [[eval [java.util.HashMap String] java.util.Map]])"
                      {:linters {:unresolved-symbol {:level :error}}})))
   (is (empty? (lint! "
+(exists? foo.bar/baz)"
+                     {:linters {:unresolved-namespace {:level :error}}}
+                     "--lang" "cljs"))))
+
+(deftest extend-type-specify-test
+  (assert-submaps
+   '({:file "<stdin>", :row 2, :col 25, :level :error, :message "Unresolved symbol: x"})
+   (lint! "
 (specify! #js {:current x}
   IDeref
   (-deref [this]
     (.-current ^js this)))"
-                     {:linters {:unresolved-symbol {:level :error}}}
-                     "--lang" "cljs"))))
+          {:linters {:unresolved-symbol {:level :error}}}
+          "--lang" "cljs"))
+  (is (empty?
+       (lint! "
+(extend-type String
+  clojure.lang.IDeref
+  (deref [this]
+    this))"
+              {:linters {:unresolved-symbol {:level :error}}}))))
 
 (deftest js-property-access-test
   (assert-submaps